Add files via upload
This program takes in user input and moves a character based on that user input.  In this example I will use the _kbhit() function.  I will use a while loop which is common in gaming.  A While Loop keeps the program/game running until a certain condition is met that will close the program.